Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/navigation-timing/nav2_test_redirect_server-expected.txt

CONSOLE MESSAGE: line 24: TypeError: undefined is not an object (evaluating 'navigation_entry.type')
Description

This test validates the values of the window.performance.redirectCount and the window.performance.timing.redirectStart/End times for a same-origin server side redirect navigation.



FAIL Navigation Timing 2 WPT TypeError: undefined is not an object (evaluating 'navigation_entry.type')
